requirements.txt
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.1
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.27.0
google-api-python-client==2.185.0
google-auth==2.41.1
google-auth-httplib2==0.2.0
google-genai==1.46.0
googleapis-common-protos==1.71.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
MarkupSafe==3.0.3
narwhals==2.9.0
numpy==2.3.4
packaging==25.0
pandas==2.3.3
pdf2image==1.17.0
pefile==2023.2.7
pillow==11.3.0
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyparsing==3.2.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
pywin32-ctypes==0.2.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
websockets==15.0.1
wheel==0.45.1
